внесла функции отрисовки и utils в модули
