+++ src/utils.js
@@ -1,0 +1,1 @@
+export const getRandomInteger = (min = 1, max = 10) => Math.floor(Math.random() * (max - min + 1)) + min;
